+++ apps/src/aichat/types/index.ts
@@ -1,5 +1,4 @@
 export * from './chatEvents';
 export * from './customizations';
 export * from './levelProperties';
-export * from './miscellaneous';
 export * from './toxicity';
