+++ apps/src/aichat/types/toxicity.ts
@@ -1,3 +1,6 @@
+import {ValueOf} from '@cdo/apps/types/utils';
+import {AiChatTeacherFeedback} from '@cdo/generated-scripts/sharedConstants';
+
 import {FIELDS_CHECKED_FOR_TOXICITY} from '../views/modelCustomization/constants';
 
 /** Response structure for the detect toxicity API */
@@ -35,3 +38,6 @@ interface ComprehendDetails {
     score: number;
   };
 }
+
+// Convenience type for the values of AiChatTeacherFeedback
+export type FeedbackValue = ValueOf<typeof AiChatTeacherFeedback>;
